# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MIB Studio Local Daemon",
    "version": "0.0.0-day0"
  },
  "components": {
    "parameters": {
      "ProjectId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/Id"
        }
      },
...
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 50
        }
      },
...
    },
    "schemas": {
      "Id": {
        "type": "string",
        "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
      },
...
      "IsoDatetime": {
        "type": "string"
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error_code",
          "message",
          "details",
          "trace_id"
        ],
        "properties": {
          "error_code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          },
          "trace_id": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "RouteInput": {
        "type": "object",
        "required": [
          "route_id",
          "description",
          "is_unsafe"
        ],
        "properties": {
          "route_id": {
            "type": "string",
            "pattern": "^[a-z0-9_]+$",
            "maxLength": 64
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "is_unsafe": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ProjectCreate": {
        "type": "object",
        "required": [
          "name",
          "preset_id",
          "routes"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "preset_id": {
            "type": "string",
            "default": "router.basic.v1"
          },
          "routes": {
            "type": "array",
            "minItems": 2,
            "maxItems": 12,
            "items": {
              "$ref": "#/components/schemas/RouteInput"
            }
          }
        },
        "additionalProperties": false
      },
      "ProjectPatch": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "routes": {
            "type": "array",
            "minItems": 2,
            "maxItems": 12,
            "items": {
              "$ref": "#/components/schemas/RouteInput"
            }
          }
        },
        "additionalProperties": false
      },
      "ProjectRead": {
        "type": "object",
        "required": [
          "id",
          "name",
          "preset_id",
          "routes",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "name": {
            "type": "string"
          },
          "preset_id": {
            "type": "string"
          },
          "routes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RouteRead"
            }
          },
          "archived_at": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IsoDatetime"
              },
              {
                "type": "null"
              }
            ]
          },
          "route_taxonomy_locked": {
            "type": "boolean",
            "default": false
          },
          "created_at": {
            "$ref": "#/components/schemas/IsoDatetime"
          },
          "updated_at": {
            "$ref": "#/components/schemas/IsoDatetime"
          }
        },
        "additionalProperties": false
      },
...
      "RouteRead": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "route_id",
          "description",
          "is_unsafe",
          "created_at"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "route_id": {
            "type": "string",
            "pattern": "^[a-z0-9_]+$",
            "maxLength": 64
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "is_unsafe": {
            "type": "boolean"
          },
          "created_at": {
            "$ref": "#/components/schemas/IsoDatetime"
          }
        }
      },
...
    },
    "responses": {
      "Error": {
        "description": "Standard error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NoContent": {
        "description": "No content"
      },
...
    }
  },
  "paths": {
...
    "/projects": {
      "get": {
        "operationId": "listProjects",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project page",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProjectRead"
                      }
                    },
                    "next_cursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createProject",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/projects/{id}": {
      "get": {
        "operationId": "getProject",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateProject",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "archiveProject",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectId"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          }
        }
      }
    },
...
  }
}